Convert winston levels to syslog levels
Instead of forcing people to install new levels and change the code, the transport itself should do as much as possible to do the automatic conversion and be compatible with winston's transport interface. I can't image the world where transport for logging system would dictate the top-level changes. Look at at `loggin` for python or `log4j` for Java.