Handle logging messages after closing the logger, and fix order
On process shutdown, one should give intent that the logger is to be shutdown.

However, log entries might still occur after that.

This change will reopen the connection, send the messages, and re-close the connection when this occurs.

While it may be overall ineffecient, this is during shutdown of the process, so it doesn't matter.

Additionally, if there are any items in the queue by the time a log entry comes in,
we need to write those queue items too, in order to maintain the order of the log entries.