Ensure winston version is >= 3.0.0; Add note in README.md too. (#102)
This bit me when I was upgrading. There was no clear documentation that this was
required and using it with 2.x.x yeilded some errors that eventually led me to
this realization. So lets make it easier for those after me who make the same
mistake by adding a nice sanity check and a note in the README.md file.