Remove unused dependency "cycle"
"cycle" is unused and appears to be unmaintained with no updates the
last six years. It does not have a license specified in package.json
so projects that use license checkers have a problem using
winston-syslog due to the presence of cycle.

The "cycle" module has ignored issues and pull requests to address the
missing license issue.

* https://github.com/dscape/cycle/issues/21
* https://github.com/dscape/cycle/pull/26